liveShareAPP/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0
django-redis==4.11.0
django-reversion==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
emoji==0.5.4
enum34==1.1.6
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.18.2
httplib2==0.9.2
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
jdcal==1.4.1
kombu==4.6.7
MarkupPy==1.14
more-itertools==8.0.2
mutagen==1.43.0
numpy==1.18.1
odfpy==1.4.0
openpyxl==3.0.2
Pillow==7.0.0
protobuf==3.11.2
pycparser==2.19
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablestore==5.1.0
tablib==0.14.0
urllib3==1.25.7
vine==1.3.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
